refactor(agent): 公共工具方法抽取到 utils/
把域无关的工具性方法从各业务模块抽到 utils/ 目录，按职责分文件 + 桶导出：
- utils/json.ts：extractJson / salvageProse / stripTrailingJson（+ 内部 escapeRawControlInStrings），
  LLM 文本的 JSON 提取 / 修复 / 打捞（原在 orchestrator.ts）。
- utils/text.ts：fillTemplate（原 prompts.ts）+ clamp（原 steps/planning/shared.ts）。
- utils/stagger.ts：runStaggered（原 src/stagger.ts，删除原文件）。
- utils/index.ts：桶导出。

各 step / autopilot-judge / orchestrator / planner 改从 utils 引用；orchestrator 退为评审驱动 + 标签 +
类型（不再夹带 JSON 工具），prompts.ts 仅留 PROMPT_TEMPLATES。index 公共导出的 extractJson 改走 utils，
公共 API 不变。纯结构调整、行为不变；agent 49 tests / lint / typecheck / desktop build 全绿。
