+++ packages/agent/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { extractJson, salvageProse, stripTrailingJson } from './json.js';
+export { fillTemplate, clamp } from './text.js';
+export { runStaggered } from './stagger.js';
